@@ -1,9 +1,9 @@
 from datetime import datetime, timedelta
-from enum import Enum
 from typing import Any, Dict, List, Optional, Tuple, Union
 from uuid import uuid4
 
 from flask import current_app
+from strenum import StrEnum
 
 from alerta.app import db
 from alerta.database.base import Query
@@ -13,7 +13,7 @@ from alerta.utils.response import absolute_url
 JSON = Dict[str, Any]
 
 
-class BlackoutStatus(str, Enum):
+class BlackoutStatus(StrEnum):
 
     Pending = 'pending'
     Active = 'active'
