@@ -147,7 +147,8 @@ def health_check():
                 return f'HEARTBEAT_STALE: {heartbeat.origin}', 503
 
     except Exception as e:
-        return f'HEALTH_CHECK_FAILED: {e}', 503
+        current_app.logger.exception(e)
+        return 'HEALTH_CHECK_FAILED: Internal Error!', 503
 
     return 'OK'
 
@@ -174,8 +175,12 @@ def housekeeping():
     errors = []
     for alert in has_expired:
         try:
-            alert = alert.from_expired(text, timeout)
+            # pre actioon
             alert, _, text, timeout = process_action(alert, action='expired', text='', timeout=None)
+            # update status
+            alert = alert.from_expired(text, timeout)
+            # post action
+            alert, _, text, timeout = process_action(alert, action='expired', text=text, timeout=timeout, post_action=True)
         except RejectException as e:
             write_audit_trail.send(current_app._get_current_object(), event='alert-expire-rejected', message=alert.text,
                                    user=g.login, customers=g.customers, scopes=g.scopes, resource_id=alert.id, type='alert',
@@ -190,8 +195,12 @@ def housekeeping():
 
     for alert in shelve_timeout + ack_timeout:
         try:
-            alert = alert.from_timeout(text, timeout)
+            # pre action
             alert, _, text, timeout = process_action(alert, action='timeout', text='', timeout=None)
+            # update status
+            alert = alert.from_timeout(text, timeout)
+            # post action
+            alert, _, text, timeout = process_action(alert, action='timeout', text=text, timeout=timeout, post_action=True)
         except RejectException as e:
             write_audit_trail.send(current_app._get_current_object(), event='alert-timeout-rejected', message=alert.text,
                                    user=g.login, customers=g.customers, scopes=g.scopes, resource_id=alert.id, type='alert',
