@@ -1,7 +1,7 @@
 import unittest
 
 from flask_cors import CORS
-from flask_cors.extension import ACL_ALLOW_HEADERS, ACL_ORIGIN
+from flask_cors.core import ACL_ALLOW_HEADERS, ACL_ORIGIN
 
 from alerta.app import create_app, db, plugins
 
